Do file checks on first upload common request (extensions and antivirus) + add double extensions check
