+++ vertigo-bot-designer/src/main/java/io/vertigo/chatbot/designer/builder/services/bot/AttachmentServices.java
@@ -1,5 +1,12 @@
 package io.vertigo.chatbot.designer.builder.services.bot;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Base64;
+import java.util.Optional;
+
+import javax.inject.Inject;
+
 import io.vertigo.account.authorization.annotations.SecuredOperation;
 import io.vertigo.chatbot.commons.LogsUtils;
 import io.vertigo.chatbot.commons.dao.AttachmentDAO;
@@ -8,57 +15,28 @@ import io.vertigo.chatbot.commons.domain.AttachmentExport;
 import io.vertigo.chatbot.commons.domain.AttachmentFileInfo;
 import io.vertigo.chatbot.commons.domain.Chatbot;
 import io.vertigo.chatbot.commons.multilingual.attachment.AttachmentMultilingualResources;
-import io.vertigo.chatbot.designer.builder.services.AntivirusServices;
 import io.vertigo.chatbot.designer.commons.services.FileServices;
 import io.vertigo.chatbot.domain.DtDefinitions;
 import io.vertigo.commons.transaction.Transactional;
 import io.vertigo.core.lang.VUserException;
-import io.vertigo.core.node.component.Activeable;
 import io.vertigo.core.node.component.Component;
-import io.vertigo.core.param.Param;
-import io.vertigo.core.param.ParamManager;
 import io.vertigo.datamodel.criteria.Criterions;
 import io.vertigo.datamodel.structure.model.DtList;
 import io.vertigo.datamodel.structure.model.DtListState;
 import io.vertigo.datamodel.structure.util.VCollectors;
 import io.vertigo.datastore.filestore.model.FileInfoURI;
 import io.vertigo.datastore.filestore.model.VFile;
-import xyz.capybara.clamav.commands.scan.result.ScanResult;
-
-import javax.inject.Inject;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Arrays;
-import java.util.Base64;
-import java.util.Collection;
-import java.util.Map;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 import static io.vertigo.chatbot.designer.utils.ListUtils.MAX_ELEMENTS_PLUS_ONE;
 
 @Transactional
-public class AttachmentServices implements Component, Activeable {
+public class AttachmentServices implements Component {
 
 	@Inject
 	private AttachmentDAO attachmentDAO;
 
 	@Inject
 	private FileServices fileServices;
-
-	@Inject
-	private AntivirusServices antivirusServices;
-
-	@Inject
-	private ParamManager paramManager;
-
-	private String[] extensionsWhiteList;
-
-	@Override
-	public void start() {
-		extensionsWhiteList = paramManager.getOptionalParam("EXTENSIONS_WHITELIST")
-				.map(Param::getValueAsString).orElse("png,jpg,jpeg,pdf").split(",");
-	}
 
 	public Attachment findById(final long attachmentId) {
 		return attachmentDAO.get(attachmentId);
@@ -79,23 +57,6 @@ public class AttachmentServices implements Component, Activeable {
 				oldAttachmentFileId = oldAttachment.getAttFiId();
 			}
 			final VFile newAttachmentFile = fileServices.getFileTmp(optFileInfoURI.get());
-			final String newAttachmentFileMimeType = newAttachmentFile.getMimeType();
-			if (Arrays.stream(extensionsWhiteList).noneMatch(newAttachmentFileMimeType::contains)) {
-				throw new VUserException(AttachmentMultilingualResources.EXTENSION_NOT_ALLOWED, newAttachmentFileMimeType,
-						String.join(",", extensionsWhiteList));
-			}
-
-			try {
-				final ScanResult result = antivirusServices.checkForViruses(newAttachmentFile.createInputStream());
-				if (result instanceof ScanResult.VirusFound) {
-					final Map<String, Collection<String>> virusesMap = ((ScanResult.VirusFound) result).getFoundViruses();
-					final String viruses = virusesMap.values().stream()
-							.flatMap(Collection::stream).collect(Collectors.joining(","));
-					throw new VUserException(AttachmentMultilingualResources.VIRUSES_FOUND, viruses);
-				}
-			} catch (final IOException ioException) {
-				throw new VUserException(AttachmentMultilingualResources.COULD_NOT_OPEN_FILE, newAttachmentFile.getFileName(), ioException);
-			}
 
 			if (isMaxSizeExceeded(maxSize, attachmentTotalSize, oldAttachmentSize, newAttachmentFile.getLength())) {
 				throw new VUserException(AttachmentMultilingualResources.MAX_TOTAL_SIZE_EXCEEDED, maxSize);
@@ -153,11 +114,6 @@ public class AttachmentServices implements Component, Activeable {
 			LogsUtils.addLogs(logs, e);
 			throw new VUserException(AttachmentMultilingualResources.EXPORT_UNEXPECTED_ERROR, e);
 		}
-
-	}
-
-	@Override
-	public void stop() {
 
 	}
 }
