+++ vertigo-bot-designer/src/main/java/io/vertigo/chatbot/designer/commons/services/FileServices.java
@@ -23,27 +23,7 @@ import com.opencsv.CSVReaderBuilder;
 import com.opencsv.bean.ColumnPositionMappingStrategy;
 import com.opencsv.bean.CsvToBean;
 import com.opencsv.exceptions.CsvValidationException;
-import io.vertigo.account.authorization.annotations.SecuredOperation;
-import io.vertigo.chatbot.commons.AttachmentInfo;
-import io.vertigo.chatbot.commons.dao.MediaFileInfoDAO;
-import io.vertigo.chatbot.commons.domain.Chatbot;
-import io.vertigo.chatbot.commons.domain.MediaFileInfo;
-import io.vertigo.chatbot.commons.multilingual.export.ExportMultilingualResources;
-import io.vertigo.commons.transaction.Transactional;
-import io.vertigo.core.lang.Assertion;
-import io.vertigo.core.lang.VSystemException;
-import io.vertigo.core.lang.VUserException;
-import io.vertigo.core.locale.LocaleManager;
-import io.vertigo.core.node.component.Component;
-import io.vertigo.datastore.filestore.FileStoreManager;
-import io.vertigo.datastore.filestore.definitions.FileInfoDefinition;
-import io.vertigo.datastore.filestore.model.FileInfo;
-import io.vertigo.datastore.filestore.model.FileInfoURI;
-import io.vertigo.datastore.filestore.model.VFile;
-import io.vertigo.datastore.filestore.util.VFileUtil;
-import io.vertigo.datastore.impl.filestore.model.StreamFile;
 
-import javax.inject.Inject;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.FileReader;
@@ -51,16 +31,48 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.Charset;
 import java.time.Instant;
+import java.util.Arrays;
 import java.util.Base64;
+import java.util.Collection;
 import java.util.List;
+import java.util.Map;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
+
+import javax.inject.Inject;
+
+import io.vertigo.account.authorization.annotations.SecuredOperation;
+import io.vertigo.chatbot.commons.AttachmentInfo;
+import io.vertigo.chatbot.commons.dao.MediaFileInfoDAO;
+import io.vertigo.chatbot.commons.domain.Chatbot;
+import io.vertigo.chatbot.commons.domain.MediaFileInfo;
+import io.vertigo.chatbot.commons.multilingual.attachment.AttachmentMultilingualResources;
+import io.vertigo.chatbot.commons.multilingual.export.ExportMultilingualResources;
+import io.vertigo.commons.transaction.Transactional;
+import io.vertigo.core.lang.Assertion;
+import io.vertigo.core.lang.VSystemException;
+import io.vertigo.core.lang.VUserException;
+import io.vertigo.core.locale.LocaleManager;
+import io.vertigo.core.node.component.Activeable;
+import io.vertigo.core.node.component.Component;
+import io.vertigo.core.param.Param;
+import io.vertigo.core.param.ParamManager;
+import io.vertigo.datastore.filestore.FileStoreManager;
+import io.vertigo.datastore.filestore.definitions.FileInfoDefinition;
+import io.vertigo.datastore.filestore.model.FileInfo;
+import io.vertigo.datastore.filestore.model.FileInfoURI;
+import io.vertigo.datastore.filestore.model.VFile;
+import io.vertigo.datastore.filestore.util.VFileUtil;
+import io.vertigo.datastore.impl.filestore.model.StreamFile;
+import xyz.capybara.clamav.commands.scan.result.ScanResult;
 
 import static io.vertigo.chatbot.designer.utils.StringUtils.lineError;
 
 @Transactional
-public class FileServices implements Component {
+public class FileServices implements Component, Activeable {
 
 	@Inject
 	private FileStoreManager fileStoreManager;
@@ -71,9 +83,48 @@ public class FileServices implements Component {
 	@Inject
 	protected LocaleManager localeManager;
 
+	@Inject
+	private ParamManager paramManager;
+
+	@Inject
+	private AntivirusServices antivirusServices;
+
+	private String[] extensionsWhiteList;
+
+	@Override
+	public void start() {
+		extensionsWhiteList = paramManager.getOptionalParam("EXTENSIONS_WHITELIST")
+				.map(Param::getValueAsString).orElse("png,jpg,jpeg,pdf,csv").split(",");
+	}
+
+	@Override
+	public void stop() {
+
+	}
+
 	public FileInfoURI saveFileTmp(final VFile file) {
 		final FileInfo fileInfo = fileStoreManager.create(new FileInfoTmp(file));
 		return fileInfo.getURI();
+	}
+
+	public void checkFile (final VFile file) {
+		final String[] extensionsTab = file.getFileName().split(Pattern.quote("."));
+		final Stream<String> extensionsWhiteListStream = Arrays.stream(extensionsWhiteList);
+		if (extensionsTab.length != 2 || extensionsWhiteListStream.noneMatch(extensionsTab[1]::contains)) {
+			throw new VUserException(AttachmentMultilingualResources.EXTENSION_NOT_ALLOWED, extensionsTab[1],
+					String.join(",", extensionsWhiteList));
+		}
+		try {
+			final ScanResult result = antivirusServices.checkForViruses(file.createInputStream());
+			if (result instanceof ScanResult.VirusFound) {
+				final Map<String, Collection<String>> virusesMap = ((ScanResult.VirusFound) result).getFoundViruses();
+				final String viruses = virusesMap.values().stream()
+						.flatMap(Collection::stream).collect(Collectors.joining(","));
+				throw new VUserException(AttachmentMultilingualResources.VIRUSES_FOUND, viruses);
+			}
+		} catch (final IOException ioException) {
+			throw new VUserException(AttachmentMultilingualResources.COULD_NOT_OPEN_FILE, file.getFileName(), ioException);
+		}
 	}
 
 	public FileInfoURI saveAttachment(final VFile file) {
